# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: TT Stock Backend API
  description: |
    A comprehensive tire and wheel inventory management system API built with Go and Gin framework.
    
    ## Features
    - User authentication and authorization with JWT tokens
    - Product management (tires and wheels)
    - Stock movement tracking
    - Low stock alerts
    - Business intelligence and reporting
    - Role-based access control (Admin, Owner, Staff)
    
    ## Authentication
    All protected endpoints require a valid JWT token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    API requests are rate limited to prevent abuse. Current limits:
    - 100 requests per minute per IP
    - 1000 requests per hour per user
    
    ## Response Times
    All API endpoints are designed to respond within 200ms under normal load conditions.
  version: 1.0.0
  contact:
    name: TT Stock API Support
    email: support@ttstock.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.ttstock.com
    description: Production server
...
paths:
...
  /api/v1/products:
    get:
      tags:
        - Products
      summary: List products
      description: Get a paginated list of products with optional filtering
      operationId: listProducts
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number (1-based)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: type
          in: query
          description: Filter by product type
          required: false
          schema:
            type: string
            enum: [Tire, Wheel]
        - name: brand
          in: query
          description: Filter by brand
          required: false
          schema:
            type: string
        - name: in_stock
          in: query
          description: Filter by stock availability
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Products
      summary: Create product
      description: Create a new product (Admin/Owner only)
      operationId: createProduct
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateRequest'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/products/{id}:
    get:
      tags:
        - Products
      summary: Get product
      description: Get a specific product by ID
      operationId: getProduct
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Products
      summary: Update product
      description: Update a product (Admin/Owner only)
      operationId: updateProduct
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdateRequest'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Products
      summary: Delete product
      description: Delete a product (Admin only)
      operationId: deleteProduct
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Product deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/products/search:
    get:
      tags:
        - Products
      summary: Search products
      description: Search products with advanced filtering and pagination
      operationId: searchProducts
      security:
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          description: Search query
          required: false
          schema:
            type: string
        - name: type
          in: query
          description: Filter by product type
          required: false
          schema:
            type: string
            enum: [Tire, Wheel]
        - name: brand
          in: query
          description: Filter by brand
          required: false
          schema:
            type: string
        - name: min_price
          in: query
          description: Minimum price filter
          required: false
          schema:
            type: number
            format: float
        - name: max_price
          in: query
          description: Maximum price filter
          required: false
          schema:
            type: number
            format: float
        - name: stock_status
          in: query
          description: Filter by stock status
          required: false
          schema:
            type: string
            enum: [available, low_stock, out_of_stock]
        - name: page
          in: query
          description: Page number (1-based)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductSearchResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/products/low-stock:
    get:
      tags:
        - Products
      summary: Get low stock products
      description: Get products with low stock levels
      operationId: getLowStockProducts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of low stock products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProductResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        type:
          type: string
          enum: [Tire, Wheel]
          example: "Tire"
        brand:
          type: string
          example: "Michelin"
        model:
          type: string
          example: "Pilot Sport 4"
        sku:
          type: string
          example: "MIC-PS4-225-45-17"
        description:
          type: string
          nullable: true
          example: "High-performance summer tire"
        cost_price:
          type: number
          format: float
          example: 120.50
        selling_price:
          type: number
          format: float
          example: 180.75
        quantity_on_hand:
          type: integer
          example: 25
        low_stock_threshold:
          type: integer
          example: 5
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    ProductCreateRequest:
      type: object
      required:
        - type
        - brand
        - model
        - sku
        - cost_price
        - selling_price
        - quantity_on_hand
        - low_stock_threshold
      properties:
        type:
          type: string
          enum: [Tire, Wheel]
          example: "Tire"
        brand:
          type: string
          example: "Michelin"
        model:
          type: string
          example: "Pilot Sport 4"
        sku:
          type: string
          example: "MIC-PS4-225-45-17"
        description:
          type: string
          nullable: true
          example: "High-performance summer tire"
        cost_price:
          type: number
          format: float
          minimum: 0
          example: 120.50
        selling_price:
          type: number
          format: float
          minimum: 0
          example: 180.75
        quantity_on_hand:
          type: integer
          minimum: 0
          example: 25
        low_stock_threshold:
          type: integer
          minimum: 0
          example: 5
    ProductUpdateRequest:
      type: object
      properties:
        type:
          type: string
          enum: [Tire, Wheel]
          example: "Tire"
        brand:
          type: string
          example: "Michelin"
        model:
          type: string
          example: "Pilot Sport 4"
        sku:
          type: string
          example: "MIC-PS4-225-45-17"
        description:
          type: string
          nullable: true
          example: "High-performance summer tire"
        cost_price:
          type: number
          format: float
          minimum: 0
          example: 120.50
        selling_price:
          type: number
          format: float
          minimum: 0
          example: 180.75
        quantity_on_hand:
          type: integer
          minimum: 0
          example: 25
        low_stock_threshold:
          type: integer
          minimum: 0
          example: 5
        is_active:
          type: boolean
          example: true
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    ProductSearchResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
        total_count:
          type: integer
          example: 150
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        total_pages:
          type: integer
          example: 8
        has_next:
          type: boolean
          example: true
        has_prev:
          type: boolean
          example: false
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Invalid request data"
        message:
          type: string
          example: "The provided data is invalid"
        details:
          type: object
          additionalProperties: true
          example:
            field: "phone_number"
            reason: "Phone number is required"
security:
  - bearerAuth: []